+++ src/protean/adapters/broker/inline.py
@@ -1,5 +1,5 @@
 # Protean
-from protean.core.domain_event import BaseDomainEvent
+from protean.core.event import BaseEvent
 from protean.globals import current_domain
 from protean.port.broker import BaseBroker
 from protean.utils import fully_qualified_name
@@ -13,7 +13,7 @@ class InlineBroker(BaseBroker):
         conn_info["IS_ASYNC"] = False
 
     def send_message(self, initiator_obj):
-        if isinstance(initiator_obj, BaseDomainEvent):
+        if isinstance(initiator_obj, BaseEvent):
             for subscriber in self._subscribers[
                 fully_qualified_name(initiator_obj.__class__)
             ]: