+++ tests/event/elements.py
@@ -2,7 +2,7 @@
 from protean.core.aggregate import BaseAggregate
 from protean.core.application_service import BaseApplicationService
 from protean.core.command import BaseCommand
-from protean.core.domain_event import BaseDomainEvent
+from protean.core.event import BaseEvent
 from protean.core.field.basic import Identifier, Integer, String
 from protean.core.unit_of_work import UnitOfWork
 from protean.globals import current_domain
@@ -34,7 +34,7 @@ class Person(BaseAggregate):
         return newcomer
 
 
-class PersonAdded(BaseDomainEvent):
+class PersonAdded(BaseEvent):
     id = Identifier(required=True)
     first_name = String(max_length=50, required=True)
     last_name = String(max_length=50, required=True)