+++ tests/subscriber/elements.py
@@ -1,6 +1,6 @@
 # Protean
 from protean.core.aggregate import BaseAggregate
-from protean.core.domain_event import BaseDomainEvent
+from protean.core.event import BaseEvent
 from protean.core.field.basic import Identifier, Integer, String
 from protean.core.subscriber import BaseSubscriber
 from protean.globals import current_domain
@@ -26,7 +26,7 @@ class Person(BaseAggregate):
         return newcomer
 
 
-class PersonAdded(BaseDomainEvent):
+class PersonAdded(BaseEvent):
     id = Identifier(required=True)
     first_name = String(max_length=50, required=True)
     last_name = String(max_length=50, required=True)
@@ -42,7 +42,7 @@ class NotifySSOSubscriber(BaseSubscriber):
     """
 
     class Meta:
-        domain_event = PersonAdded
+        event = PersonAdded
 
-    def notify(self, domain_event):
-        print("Received Domain Event: ", domain_event)
+    def notify(self, event):
+        print("Received Event: ", event)